Upgraded RN to v0.31
_Why_?

So we can do more fun and interesting things with better dependency
resolution (hopefully)

_How_?

installed stuff
ran react-native upgrade

_Side Effects_?

no.
