+++ package.json
@@ -77,8 +77,8 @@
   },
   "dependencies": {
     "lodash": "^4.13.1",
-    "react": "^15.1.0",
-    "react-native": "^0.27.2",
+    "react": "^15.2.1",
+    "react-native": "^0.31.0",
     "react-redux": "^4.4.5",
     "redux": "^3.5.2",
     "redux-logger": "^2.6.1",
